part4 4.21

+++ part4/blogList/controllers/blogs.js
@@ -42,6 +42,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+    const blog = await Blog.findById(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id || decodedToken.id.toString() !== blog.user.toString()) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
